feat: 1단계 - 예외 처리와 응답 구현
- API 요청에 대한 예외 처리
- 예외 발생 시 400 Status code 응답 처리

+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -2,17 +2,23 @@ package roomescape.service;
 
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.springframework.web.bind.annotation.ExceptionHandler;
 import roomescape.domain.ReservationTime;
 import roomescape.dto.request.ReservationTimeRequest;
+import roomescape.repository.ReservationRepository;
 import roomescape.repository.ReservationTimeRepository;
 
 @Service
 public class ReservationTimeService {
+    private final ReservationRepository reservationRepository;
     private final ReservationTimeRepository reservationTimeRepository;
 
     @Autowired
-    public ReservationTimeService(ReservationTimeRepository reservationTimeRepository) {
+    public ReservationTimeService(ReservationRepository reservationRepository, ReservationTimeRepository reservationTimeRepository) {
+        this.reservationRepository = reservationRepository;
         this.reservationTimeRepository = reservationTimeRepository;
     }
 
@@ -25,6 +31,11 @@ public class ReservationTimeService {
     }
 
     public void deleteReservationTime(Long id) {
+        boolean existsByTimeId = reservationRepository.existsByTimeId(id);
+        if (existsByTimeId) {
+            throw new IllegalArgumentException();
+        }
+
         reservationTimeRepository.deleteReservationTime(id);
     }
 }
